+++ src/extension.ts
@@ -1,34 +1,18 @@
-import * as vscode from 'vscode';
-import * as path from 'path'
-import * as fs from 'fs'
-import { SlidesPreviewPanel } from './SlidesPreviewPanel';
-import { createRevealJsHtml, showErrorMessage } from './utils';
-const slash = require("slash")
+import * as vscode from 'vscode'
+import { exportHtml } from './commands/exportHtml'
+import { exportInlinedHtml } from './commands/exportInlinedHtml'
+import { showPreview } from './commands/showPreview'
+import { openInBrowser } from './commands/openInBrowser'
+import { ContainerManager } from './ContainerManager'
 
 export function activate(context: vscode.ExtensionContext) {
 
-	context.subscriptions.push(vscode.commands.registerCommand('asciidocSlides.preview', () => {
-		SlidesPreviewPanel.createOrShow(context.extensionPath)
-	}));
+	const outputChannel = vscode.window.createOutputChannel("asciidoc slides")
+	const appendLine = (value: string) => outputChannel.appendLine(value)
+	const containerManager = new ContainerManager(context, appendLine)
 
-	context.subscriptions.push(vscode.commands.registerCommand('asciidocSlides.export', async () => {
-		const document = vscode.window.activeTextEditor?.document
-		if(document) {
-			const proposedFilename = path.join(path.dirname(document.fileName), "slides.html")
-			const exportFileLocation = await vscode.window.showSaveDialog({defaultUri: vscode.Uri.file(proposedFilename), filters: {'HTML': ['html']}})
-			if(exportFileLocation) {
-				const pathCompleter = (inputPath: string) => slash(path.join(context.extensionPath, inputPath))
-				const resourceBasePath = slash(path.dirname(document.fileName))
-				const slidesHtml = createRevealJsHtml(document.getText(), pathCompleter, resourceBasePath, false);
-				fs.writeFile(exportFileLocation.fsPath, slidesHtml, (err) => {
-					if(err) {
-						showErrorMessage(`Error while exporting: ${err?.message}`)
-					}
-				})
-			}
-
-		} else {
-			showErrorMessage("Call this command based on an asciidoc document.")
-		}
-	}));
+	context.subscriptions.push(vscode.commands.registerCommand('asciidocSlides.preview', () => showPreview(containerManager)))
+	context.subscriptions.push(vscode.commands.registerCommand('asciidocSlides.exportHtml', () => exportHtml(containerManager)))
+	context.subscriptions.push(vscode.commands.registerCommand('asciidocSlides.exportInlinedHtml', () => exportInlinedHtml(containerManager)))
+	context.subscriptions.push(vscode.commands.registerCommand('asciidocSlides.openInBrowser', () => openInBrowser(containerManager)))
 }